create separate controller for provider secret migration
