+++ Makefile.prow
@@ -12,7 +12,7 @@ compile:
 	go mod vendor
 	go mod tidy
 	GOFLAGS="" go build -o build/_output/manager ./cmd/manager/main.go
-	GOFLAGS="" go build -o build/_output/secretupdater ./cmd/secretupdater/main.go
+	GOFLAGS="" go build -o build/_output/old-provider-connection ./cmd/oldproviderconnection/main.go
 
 .PHONY: build
 build:
